Separated RideInProgressCard to its own file
